// lib/db/types/couriers/webhook.ts
export enum EventOptions {
  TrackingCreate = "tracking_create",
  TrackingUpdate = "tracking_update",
  TrackingDelete = "tracking_delete",
  TrackingCheckpointUpdate = "tracking_checkpoint_update",
  ShipmentCreate = "shipment_create",
  ShipmentUpdate = "shipment_update",
  ShipmentDelete = "shipment_delete",
  ShipmentCancel = "shipment_cancel",
  ShipmentGenerated = "shipment_generated",
}